--- exercises/PQ/Library/library/endpoint_test.go
@@ -1,1 +1,0 @@
-package library
