--- exercises/PQ/Library/library/repository_test.go
@@ -1,1 +1,0 @@
-package library
