--- exercises/PQ/Library/library/transport_test.go
@@ -1,1 +1,0 @@
-package library
