+++ exercises/PQ/Library/materials/domain.go
@@ -1,4 +1,4 @@
-package library
+package materials
 
 import "time"
