+++ exercises/PQ/Library/materials/endpoint.go
@@ -1,4 +1,4 @@
-package library
+package materials
 
 import (
 	"context"
