+++ exercises/PQ/Library/materials/errors.go
@@ -1,4 +1,4 @@
-package library
+package materials
 
 import "errors"
