+++ exercises/PQ/Library/materials/repository_mock.go
@@ -1,4 +1,4 @@
-package library
+package materials
 
 import (
 	"context"
