+++ exercises/PQ/Library/materials/service.go
@@ -1,4 +1,4 @@
-package library
+package materials
 
 import (
 	"context"
